chore: windows build fixes
