+++ pkg/command/constants.go
@@ -1,0 +1,6 @@
+package command
+
+const (
+	SHELL      = "/bin/sh"
+	POWERSHELL = "powershell"
+)
