+++ utils/helpers_test.go
@@ -1,7 +1,6 @@
 package utils
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
 	"io"
@@ -13,9 +12,7 @@ import (
 	"time"
 
 	"github.com/AlecAivazis/survey/v2"
-	"github.com/aziontech/azion-cli/pkg/cmdutil"
 	"github.com/aziontech/azion-cli/pkg/contracts"
-	"github.com/aziontech/azion-cli/pkg/iostreams"
 	"github.com/aziontech/azion-cli/pkg/logger"
 	"github.com/manifoldco/promptui"
 	"github.com/stretchr/testify/assert"
@@ -154,48 +151,6 @@ func TestIsDirEmpty(t *testing.T) {
 	}
 }
 
-func TestRunCommandWithOutput(t *testing.T) {
-	tests := []struct {
-		name       string
-		envVars    []string
-		command    string
-		wantOutput string
-		wantCode   int
-		wantErr    bool
-	}{
-		{
-			name:       "Successful command without env vars",
-			envVars:    nil,
-			command:    "echo 'Hello, World!'",
-			wantOutput: "Hello, World!\n",
-			wantCode:   0,
-			wantErr:    false,
-		},
-		{
-			name:       "Successful command with env vars",
-			envVars:    []string{"FOO=bar"},
-			command:    "echo $FOO",
-			wantOutput: "bar\n",
-			wantCode:   0,
-			wantErr:    false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			output, exitCode, err := RunCommandWithOutput(tt.envVars, tt.command)
-
-			if (err != nil) != tt.wantErr {
-				t.Errorf("RunCommandWithOutput() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			require.Equal(t, tt.wantOutput, output)
-			require.Equal(t, tt.wantCode, exitCode)
-		})
-	}
-}
-
 func TestLoadEnvVarsFromFile(t *testing.T) {
 	tests := []struct {
 		name        string
@@ -246,273 +201,6 @@ func TestLoadEnvVarsFromFile(t *testing.T) {
 
 			if tt.name != "Invalid env file path" {
 				_ = os.RemoveAll(filepath.Dir(filePath))
-			}
-		})
-	}
-}
-
-func TestCommandRunInteractiveWithOutput(t *testing.T) {
-	tests := []struct {
-		name       string
-		factory    *cmdutil.Factory
-		command    string
-		envVars    []string
-		wantOutput string
-		wantErr    bool
-	}{
-		{
-			name: "Successful command without env vars and silent false",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: false,
-					},
-				},
-			},
-			command:    "echo 'Hello, World!'",
-			envVars:    nil,
-			wantOutput: "Hello, World!\n",
-			wantErr:    false,
-		},
-		{
-			name: "Successful command with env vars and silent false",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: false,
-					},
-				},
-			},
-			command:    "echo $FOO",
-			envVars:    []string{"FOO=bar"},
-			wantOutput: "bar\n",
-			wantErr:    false,
-		},
-		{
-			name: "Failed command without env vars and silent false",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: false,
-					},
-				},
-			},
-			command:    "invalidcommand",
-			envVars:    nil,
-			wantOutput: "",
-			wantErr:    true,
-		},
-		{
-			name: "Failed command with env vars and silent false",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: false,
-					},
-				},
-			},
-			command:    "invalidcommand",
-			envVars:    []string{"FOO=bar"},
-			wantOutput: "",
-			wantErr:    true,
-		},
-		{
-			name: "Successful command with env vars and silent true",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: true,
-					},
-				},
-			},
-			command:    "echo $FOO",
-			envVars:    []string{"FOO=bar"},
-			wantOutput: "",
-			wantErr:    false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			output, err := CommandRunInteractiveWithOutput(tt.factory, tt.command, tt.envVars)
-
-			if (err != nil) != tt.wantErr {
-				t.Errorf("CommandRunInteractiveWithOutput() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			require.Equal(t, tt.wantOutput, output)
-		})
-	}
-}
-
-func TestCommandRunInteractive(t *testing.T) {
-	tests := []struct {
-		name    string
-		factory *cmdutil.Factory
-		command string
-		wantErr bool
-	}{
-		{
-			name: "Successful command with Silent false and no Format or Out",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: false,
-					},
-				},
-			},
-			command: "echo 'Hello, World!'",
-			wantErr: false,
-		},
-		{
-			name: "Successful command with Silent true",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: true,
-					},
-				},
-			},
-			command: "echo 'Hello, World!'",
-			wantErr: false,
-		},
-		{
-			name: "Successful command with Format flag",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: false,
-					},
-					Format: "json",
-				},
-			},
-			command: "echo 'Hello, World!'",
-			wantErr: false,
-		},
-		{
-			name: "Successful command with Out flag",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: false,
-					},
-					Out: "output.txt",
-				},
-			},
-			command: "echo 'Hello, World!'",
-			wantErr: false,
-		},
-		{
-			name: "Failed command",
-			factory: &cmdutil.Factory{
-				IOStreams: &iostreams.IOStreams{
-					In:  os.Stdin,
-					Out: os.Stdout,
-					Err: os.Stderr,
-				},
-				Flags: cmdutil.Flags{
-					Logger: logger.Logger{
-						Silent: false,
-					},
-					Out: "output.txt",
-				},
-			},
-			command: "invalidcommand",
-			wantErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := CommandRunInteractive(tt.factory, tt.command)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("CommandRunInteractive() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func TestRunCommandStreamOutput(t *testing.T) {
-	tests := []struct {
-		name       string
-		envVars    []string
-		comm       string
-		wantOutput string
-		wantErr    bool
-	}{
-		{
-			name:       "Successful command without env vars",
-			envVars:    nil,
-			comm:       "echo 'Hello, World!'",
-			wantOutput: "Hello, World!\n",
-			wantErr:    false,
-		},
-		{
-			name:       "Successful command with env vars",
-			envVars:    []string{"FOO=bar"},
-			comm:       "echo $FOO",
-			wantOutput: "bar\n",
-			wantErr:    false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			var out bytes.Buffer
-			err := RunCommandStreamOutput(&out, tt.envVars, tt.comm)
-
-			if (err != nil) != tt.wantErr {
-				t.Errorf("RunCommandStreamOutput() error = %v, wantErr %v", err, tt.wantErr)
-			}
-
-			if gotOutput := out.String(); gotOutput != tt.wantOutput {
-				t.Errorf("RunCommandStreamOutput() = %v, want %v", gotOutput, tt.wantOutput)
 			}
 		})
 	}
